// packages/trading-bot-server/models/KrakenOpenPosition.ts
export interface KrakenOpenPosition {
  cost: string;
  fee: string;
  margin: string;
  misc: string;
  oflags: string;
  ordertxid: string;
  ordertype: string;
  pair: string;
  posstatus: string;
  rollovertm: string;
  terms: string;
  time: number;
  type: string;
  vol: string;
  vol_closed: string;
}